requirements.txt
feedparser
requests
python-dotenv
# Optional: include model client if you use a specific SDK
